Revise vim installation
For ubuntu 16.04 and 18.04, vim is just bit old does not be supported
from the latest useful tools like as dein. Neovim was one of the best
choices for, but the latest ver of vim also become another choice.

This update is to add installation for the latest vim and enable to
select which of vim is installed.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>