Add support for empty input for config task
For setting up YAML files under `group_vars` dir, the installer define
value of hash entry as "None" if user input is empty. Howevrer, it
should be actual empty value because "None" is ambiguous for empty or
string value "None" while parsing the YAML file later.

This update is to change the string value to empty for the issue.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>